Add support for cloud-init on LXC instances
Images that use cloud-init are not correctly initialized when using
libvirt LXC nova driver.

One way cloud-init checks if the OpenStack datasource should be used
is by checking DMI data that is meaningful for virtual machines but
not for containers.

Another way cloud-init is using is to check if the 'product_name'
env variable for init process (PID 1) is "OpenStack Nova" [1][2].

This commit add that env variable to the instance when the driver
is LXC.

[1] https://cloudinit.readthedocs.io/en/latest/topics/datasources/openstack.html
[2] https://git.launchpad.net/cloud-init/tree/tools/ds-identify#n974

Closes-Bug: 1834506

Change-Id: I2d0a4461081f5284d16df73a783cb7dae3ff0ef5
Signed-off-by: Miguel Herranz <miguel@midokura.com>