Add integration testing for heal_allocations
This adds a simple scenario for the heal_allocations CLI
to the post_test_hook script run at the end of the nova-next
job. The functional testing in-tree is pretty extensive but
it's always good to have real integration testing.

NOTE(melwitt): This is different than the Train change and does not
use the --dry-run or --instance options because the following changes
are not in Stein:

  Ide31957306602c1f306ebfa48d6e95f48b1e8ead
  Icf57f217f03ac52b1443addc34aa5128661a8554

Change-Id: If86e4796a9db3020d4fdb751e8bc771c6f98aa47
Related-Bug: #1819923
(cherry picked from commit 87365c760ee89493db87183c4c751ba5fc7b921a)