Get rid of args to RBDDriver.__init__()
While reviewing [1] I noticed an annoying and unnecessary
coupling/duplication of references to the CONF variables for RBD. This
commit consolidates them down to a single source, the guts of
RBDDriver.__init__().

As part of this a workaround is also removed concerning the passing of
conffile=None to RADOS via python-rbd. The following change merged in
2014 so we can now return None when ``[libvirt]/images_rbd_ceph_conf``
is not set:

Make python rados.Rados accept conffile=None
https://github.com/ceph/ceph/pull/1787

[1] https://review.opendev.org/667421

Co-Authored-By: Lee Yarwood <lyarwood@redhat.com>
Change-Id: I361af845d6a733618ecd056aa7df973191184ae9