Set COMPUTE_STATUS_DISABLED trait from update_provider_tree flow
This adds code which hooks in the update_provider_tree flow in the
ResourceTracker, specifically when the RT is generically modifying
the compute-owned traits for the given compute node resource provider.

A future change will add the scheduler request pre-filter and
API code to sync the trait when enabling/disabling a compute service.
This is necessary for two cases specifically:

1. After upgrading an older disabled compute we will sync the trait.

2. If enabling/disabling a compute and the service is down, the API
   will not call the compute service to sync the trait. When the
   compute service is restarted we will sync the trait on startup
   with this code.

The COMPUTE_STATUS_DISABLED trait was added to os-traits in change
Ia8e4487bfb59f764a6817ec8650785ffa902eab5 which is in the 0.15.0
release of os-traits so the requirements are bumped here as well.

Part of blueprint pre-filter-disabled-computes

Change-Id: I3005b46221ac3c0e559e1072131a7e4846c9867c