Add VirtAPI.update_compute_provider_status
This adds the VirtAPI interface to add/remove the
COMPUTE_STATUS_DISABLED trait on a compute node
resource provider based on whether or not the related
compute service is disabled. This will be used in future
patches by both the ComputeManager's set_host_enabled method,
which will be called from the API when a compute service is
enabled/disabled, and from the libvirt driver when the
hypervisor connection is dropped/re-connected.

Note that some refactoring was required to make sure the
ComputeManager, ResourceTracker and ComputeVirtAPI are all
using the same SchedulerReportClient instance with the same
ProviderTree cache since the set_traits_for_provider method
relies on the cache to determine if updates are flushed back
to Placement.

Part of blueprint pre-filter-disabled-computes

Change-Id: Ie756ba5e405ad988667c75a723f1c9b9ff3e4a93