libvirt: manage COMPUTE_STATUS_DISABLED for hypervisor connection
The libvirt driver has a callback method for when the hypervisor
connection drops and is re-connected which mirrors those changes
in the related compute Service record's "disabled" field.

This builds on that by hooking back into the ComputeVirtAPI
update_compute_provider_status method when the Service.disabled
value changes so the COMPUTE_STATUS_DISABLED trait can be added
to or removed from the related compute node resource provider.

Part of blueprint pre-filter-disabled-computes

Change-Id: Ifabbb543aab62b917394eefe48126231df7cd503