Refactor HostAPI.service_update
Change If32bca070185937ef83f689b7163d965a89ec10a is going to
reflect a compute service's disabled status with the
COMPUTE_STATUS_DISABLED trait in placement. This change
does some prep work so that a common method in the HostAPI
(service_update) can be used to abstract that for pre-2.53
and post-2.53 service update APIs.

In the pre-2.53 case, the Service is retrieved using the
host and binary which requires looking up the HostMapping.
In the post-2.53 case, we already have the Service object and
just need to save the changes.

Part of blueprint pre-filter-disabled-computes

Change-Id: Id916ccd21542c90bf80caec246dfdba72a9c58b8