libvirt: Remove unnecessary try-catch around 'getCPUMap'
This was added in the original review [1] but seems to be totally
unnecessary: the only reason this can fail is if libvirt version is less
than 1.2.5 [2] or if the ACL checks fail [3]. The former isn't an issue
since our minimum is greater than this and the latter isn't an issue
since that feature is off by default and there's no reason it would be
turned on for a nova host, since nova is a trusted management
application that needs full access to libvirt.

Remove the exception handling and fix a bug with a related exception,
which was attempting to do delayed string interpolation which is only a
thing with the 'logging' library, not exceptions.

[1] https://review.opendev.org/#/c/173187/8/nova/virt/libvirt/driver.py@4480
[2] https://github.com/libvirt/libvirt/search?l=C&q=.nodeGetCPUMap
[3] https://github.com/libvirt/libvirt/blob/76be4f5ddac/src/vz/vz_driver.c#L1003

Change-Id: Ia7bfd3e9c8b542beb59a7a01da45a60a6027ea37
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>