Refresh instance in MigrationTask.execute Exception handler
Cross-cell resize is a synchronous operation from the MigrationTask
which means if something fails on the compute and sets the instance
to ERROR, the Exception block in _cold_migrate in conductor can
blindly reset the vm_state to ACTIVE because it is using a stale
copy of the instance.

This change refreshes the instance before calling _set_vm_state_and_notify
so we don't overwrite the current vm_state. This is not really a
problem for same-cell resize because prep_resize is an RPC cast to
the destination compute so any failures there will be persisted
separately from the Exception block in the _cold_migrate method.

Change-Id: I2254e36407c9f9ff674eaec44c115d7516421de3