Pass migration to finish_revert_migration()
In order to obtain plug-time events, the libvirt driver needs to pass
a Migration object to the network model. Previously [1], this object
was obtained with a database lookup. This was done to avoid modifying
the virt driver interface, thus making the fix backportable. This
patch makes the compute manager pass the migration object to the virt
driver in order to avoid a needless database query.

[1] https://review.opendev.org/#/c/667177/

Change-Id: I9095c1be65b127d40f5bbe7af5d63fefedb73d33