Remove Newton-era min compute checks for server create with device tags
Compute service version 14 is from Newton [1] so we can remove
the API min compute version checks for creating a server with
bdm and vif device tags. Note that the docstrings which are
removed mentioned microversion support but that was wrong - the
request schema validation is what validates the microversion used.

[1] I8367f740d6d4ebaeb81bc74c6a82a8faf5cd8308

Change-Id: I97e67fb971b7a0cc2373b558907c7354646cf5fa