Nova compute: add in log exception to help debug failures
When the system is under load, there can be boot failures. One
exception seen in the log file is:

''NoneType' object has no attribute 'status_code''

There is no log trace which helps pinpoint this client call.
Note that this is only one example.

We now add in a LOG.exception to have the stack trace.

Related-Bug: #1850682

Change-Id: I2ab56005db363f81eecf07c3db62c267ba7f8ca9