Add compatibility checks for CPU mode and CPU models and extra flags
Add new method '_check_cpu_compability' in driver. It is used to check
if the configuration of cpu_mode/cpu_models/cpu_model_extra_flags is
correct and compatible with guest CPU.

The method will be called when the service(use libvirt driver) start, if
the configuration is not correct, an exception will be raised.

Change-Id: Ib37ff6fc2e1f9de0e60adca54e87c74f45f12ffa
Implements: blueprint cpu-model-selection