Avoid logging traceback when detach device not found
We use the oslo.utils save_and_reraise_exception context manager in our
detach device code and catch specific exceptions that mean 'not found'
and raise DeviceNotFound instead. When we do that, the
save_and_reraise_exception context manager logs an ERROR traceback of
the original exception, for informational purposes. This is misleading
when trying to debug other issues, as it makes it look like the caught
exception caused a problem.

This passes the reraise=False keyword arg to the context manager and
sets the 'reraise' attribute to True only if we are not going to raise
a different exception.

Related-Bug: #1836212

Change-Id: Icce1e31fe3ebcbf9e4897bbfa57b7f3d1fba67a3
(cherry picked from commit 738774b942dab21d8c03c70b4e74b6a487515ff5)