Poison netifaces.interfaces() in tests
While reviewing [1] it was noted that we've probably been invoking the
real netifaces.interfaces() in some test paths. This isn't good, so
introduce a fixture that blows up any test that does that.

(NB: during development, the netifaces.interfaces()-specific poison
fixture evolved into a generic fixture that can be extended in the
future to poison anything in a similar fashion.)

[1] https://review.opendev.org/#/c/671471/

Change-Id: I1fea14d5be10bb4e884f52e0ae8be722519ddd3f