Rework 'hardware.numa_usage_from_instances'
This function was confusing. Its name suggests it is expecting multiple
instances from which to parse usage, when in reality it was only ever
called with a single instance NUMA topology from non-test code. Rename
and rework the function, using a better function name, a better
docstring and normalized variable names both in this function and two
related functions, in order to make the whole thing a little more sane.
Any tests that relied on the ability to pass multiple instance NUMA
topology objects in are reworked to simply call the function multiple
times.

Change-Id: Ic6f5f89ffaaff1bdc213c1d2b8b56aad99bdd594
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>