Add functional regression test for bug 1837955
This adds a functional regression recreate test for
bug 1837955 which was introduced with change
Iae904afb6cb4fcea8bb27741d774ffbe986a5fb4 in the Queens
release.

In this scenario, the primary (and potentially alternate)
hosts for a server build fail and reschedule to conductor.
Eventually all alternate hosts are exhausted and specifically
trying to claim allocations against the alternates fails,
probably because between the time of initial scheduling and
rescheduling something else took up the spare capacity on the
alternate host.

When this happens, MaxRetriesExceeded is raised but the
instance is stuck in BUILD status rather than set to ERROR
status with a fault message.

Change-Id: I4ca64dd60d883356880680fb1f04cee4136c2e00
Related-Bug: #1837955
(cherry picked from commit 5cc39fc51eee6e39f584394283f009aa6809d17b)