neutron: log something more useful in _get_instance_nw_info
This method is currently logging something like this:

  Jul 30 20:47:58 devstack nova-compute[23811]: DEBUG \
  nova.network.neutronv2.api [None req-1bba63ae-86d8-4a97-be63-18f18c9ae662 \
  None None] [instance: 3e0a7f65-34ac-4d73-83ed-96d36fdefd54] \
  _get_instance_nw_info() {{(pid=23811) _get_instance_nw_info \
  /opt/stack/nova/nova/network/neutronv2/api.py:1751}}

Logging the method name, which already gets logged, is not very
useful.

This change uses the parameters to the method to add some more useful
log messages.

Currently the only time force_refresh is True is when the
_heal_instance_info_cache periodic task runs.

Currently refresh_vif_id is only ever not None when a network-changed
event is received with a particular port id in the event tag parameter.

Change-Id: Iccda422a3900821ca5476d8888e935568884f321