rt: soften warning case in _remove_deleted_instances_allocations
During an instance_claim during initial server create or unshelve,
the instance.host/node values will be set before the resource
tracker has the instance in the tracked_instances dict. If
_remove_deleted_instances_allocations is called with the instance
before it's being tracked, a warning like this is logged:

  Jul 31 13:12:57.455904 ubuntu-bionic-rax-iad-0009534722
  nova-compute[31951]: WARNING nova.compute.resource_tracker
  [None req-d6f2ae97-d8f7-46f6-8974-b42aeb58302d None None]
  Instance 227c23cd-aeb2-4b5a-b001-21bd920a5e77 is not being actively
  managed by this compute host but has allocations referencing this
  compute host: {u'resources': {u'MEMORY_MB': 64, u'VCPU': 1,
  u'DISK_GB': 1}}. Skipping heal of allocation because we do not know
  what to do.

This shows up quite frequently in CI runs (see the bug report for
a logstash query) which means it should not be a warning.

This change checks the instance task_state and if set then we only
log a debug message rather than the warning since we can assume we
are racing and the task will correct itself upon completion.

Change-Id: I6db8bea6761b68c39e6332d4698d1f8312863758
Closes-Bug: #1838541
(cherry picked from commit ad4ef5af8bf7dc34caef21b0062ef9cc504bc216)