trivial: Rewrap definitions of 'NUMACell'
In a future change, I492803eaacc34c69af073689f9159449557919db, we're
going to be adding a new field to all definitions of this object. That
is going to result in a lot of line wrapping which will make the change
a lot harder to grok. Do that wrapping now so that we can make the
functional changes in that future change easier to identify.

Part of blueprint cpu-resources

Change-Id: Id3441073adde563a568c9550df53690d1e6c998a
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>