Add nova-osprofiler-redis job to experimental queue
This adds a nova-specific variant of the
tempest-smoke-py3-osprofiler-redis job to nova's
experimental queue. This could be useful in comparing
before and after results on nova changes that are
related to performance, although that's admittedly
hard in a CI environment where we could be running on
different node providers (we could pin that though?)
and differing levels of noisy neighbors during each
job run.

Change-Id: I2dcb5c1e8289eaf7977b38a3d26e74183823c42b