objects: Add 'NUMACell.pcpuset' field
We will use this to track host CPUs intended for dedicated workloads in
a future change. For now, we simply add the field and set it in all
instances of 'NUMACell'. This change has the benefit of allowing us to
avoid overloading the meaning of some fields, as described below. This
doesn't happen here but rather with change
I492803eaacc34c69af073689f9159449557919db.

Before:

  pinned instances ->

    NUMACell.cpuset
      Tracks the host cores that can be used by pinned instances
    NUMACell.cpu_usage
      Tracks the usage of host cores by pinned instances in total
      integer format
    NUMACell.pinned_cpus
      Tracks the usage of host cores by pinned instances in mapping
      format

  unpinned instances ->

    NUMACell.cpuset
      Tracks the host cores that can be used by unpinned instances
    NUMACell.cpu_usage
      Tracks the usage of host cores by unpinned instances in total
      integer format

After:

  pinned instances ->

    NUMACell.pcpuset
      Tracks the host cores that can be used by pinned instances
    NUMACell.pinned_cpus
      Tracks the usage of host cores by pinned instances in mapping
      format

  unpinned instances ->

    NUMACell.cpuset
      Tracks the host cores that can be used by unpinned instances
    NUMACell.cpu_usage
      Tracks the usage of host cores by unpinned instances in total
      integer format

In summary, we go from overloading the 'cpuset' and 'cpu_usage' fields
to having dedicated fields for pinned and unpinned CPUs. As an aside, it
would be nice if the field names were a little more uniform, but that
can be addressed in the future.

Change-Id: I7407a8e662d7b4a710d5a133ef22d2b25e26ad89
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>